neutts/pyproject.toml
[project]
name = "neutts-pi"
version = "0.1.0"
description = "Neuphonic Neutts for Raspberry Pi"
requires-python = ">=3.10, <=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "librosa>=0.11.0",
    "llama-cpp-python>=0.3.16",
    "neucodec>=0.0.4",
    "numpy>=2.2.6",
    "phonemizer>=3.3.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "resemble-perth>=1.0.1",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I", "F"]
